+++ client/src/http/Field.ts
@@ -1,16 +1,16 @@
 import { DataType } from '../pages/collections/Types';
-import { FieldView } from '../pages/structure/Types';
+import { FieldData } from '../pages/structure/Types';
 import { IndexState } from '../types/Milvus';
 import BaseModel from './BaseModel';
-import { IndexHttp } from './Index';
 
-export class FieldHttp extends BaseModel implements FieldView {
+export class FieldHttp extends BaseModel implements FieldData {
   data_type!: DataType;
   fieldID!: string;
   type_params!: { key: string; value: string }[];
   is_primary_key!: true;
   name!: string;
-  // data from index http
+
+  // index
   _indexType!: string;
   _indexParameterPairs!: { key: string; value: string }[];
   _indexStatus!: IndexState;
@@ -30,26 +30,6 @@ export class FieldHttp extends BaseModel implements FieldView {
     });
 
     return res.schema.fields.map((f: any) => new this(f));
-  }
-
-  static async getStructureListWithIndex(
-    collectionName: string
-  ): Promise<FieldHttp[]> {
-    const vectorTypes: DataType[] = ['BinaryVector', 'FloatVector'];
-    const indexList = await IndexHttp.getIndexInfo(collectionName);
-    const structureList = [...(await this.getFields(collectionName))];
-    let fields: FieldHttp[] = [];
-    for (const structure of structureList) {
-      if (vectorTypes.includes(structure.data_type)) {
-        const index = indexList.find(i => i._fieldName === structure.name);
-        structure._indexParameterPairs = index?._indexParameterPairs || [];
-        structure._indexType = index?._indexType || '';
-        structure._createIndexDisabled = indexList.length > 0;
-      }
-
-      fields = [...fields, structure];
-    }
-    return fields;
   }
 
   get _fieldId() {
